+++ neuralserv-api/src/main/java/org/reggiemcdonald/persistence/TrainingSessionRowMapper.java
