+++ neuralserv-api/src/main/java/org/reggiemcdonald/persistence/dao/TrainingSessionDao.java
@@ -1,5 +1,6 @@
 package org.reggiemcdonald.persistence.dao;
 
+import org.reggiemcdonald.exception.TrainingSessionNotFoundException;
 import org.reggiemcdonald.persistence.dto.TrainingSessionDto;
 
 import java.util.Date;
@@ -15,15 +16,16 @@ public interface TrainingSessionDao {
 
     /**
      * Get all training sessions between a particular date
-     * @param trainingDate
+     * @param startDate: The earliest date to include training sessions
+     * @param endDate: The latest date to include training sessions
      * @return
      */
-    List<TrainingSessionDto> findByDate(Date trainingDate);
+    List<TrainingSessionDto> findByDate(Date startDate, Date endDate);
 
     /**
      * Get a particular training session
      * @param id
      * @return
      */
-    TrainingSessionDto findById(int id);
+    TrainingSessionDto findById(int id) throws TrainingSessionNotFoundException;
 }
