+++ neuralserv-api/src/main/java/org/reggiemcdonald/persistence/dto/TrainingSessionDto.java
